+++ src/Symfony/Bundle/FrameworkBundle/Tests/Functional/Bundle/DebugAutoconfigurationBundle/Autoconfiguration/Bindings.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\DebugAutoconfigurationBundle\Autoconfiguration;
+
+class Bindings
+{
+    private $paramOne;
+    private $paramTwo;
+
+    public function __construct($paramOne, $paramTwo)
+    {
+        $this->paramOne = $paramOne;
+        $this->paramTwo = $paramTwo;
+    }
+}