+++ src/Symfony/Bundle/FrameworkBundle/Tests/Functional/Bundle/DebugAutoconfigurationBundle/Autoconfiguration/MethodCalls.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\DebugAutoconfigurationBundle\Autoconfiguration;
+
+class MethodCalls
+{
+    public function setMethodCallOne()
+    {
+    }
+
+    public function setMethodCallTwo()
+    {
+    }
+}