+++ src/Symfony/Bundle/FrameworkBundle/Tests/Functional/Bundle/DebugAutoconfigurationBundle/DebugAutoconfigurationBundle.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Symfony\Bundle\FrameworkBundle\Tests\Functional\Bundle\DebugAutoconfigurationBundle;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class DebugAutoconfigurationBundle extends Bundle
+{
+}